Updates representations of query components and reorganizes sql modules.
Three types of query component have been defined:
`TableExpression`, `Referenceable`, and `Selectable`.
A `TableExpression` is suitable for use in from-clauses and anywhere else
a table-like structure would appear.
A `Referenceable` can be used in most other parts of the query. These would
be the 'SELECT', 'WHERE', 'GROUP BY', 'HAVING', and 'ORDER BY' portions.
A `Selectable can be used only in a the select clause of a query. Most
referenceable items should also be selectable, but not necessarily the
other way around.

Many operators have been moved into the expressions module to unify the
two concepts, and due to the tight coupling between them and all
expressions.