Fixes the SQL `IN` operator.
In_ is the (currently) only subclass of a _SubqueryOperator, which is a binary
operator that treats the second argument like a table expression instead of
a value expression. At some point a better distinction between a table and a
query may be necessary, but this allows queries to be used in this context
without problems for the meantime.