Fixes HAVING clause code and adds unit tests for GROUP BY and HAVING
A misnamed variable in HAVING clause generation was causing an
error when adding HAVING conditions, which is now resolved.
Additionaly, trying to execute a query with a having condition but
no grouping expressions produces and error.

Unit tests cover grouping, grouping with a having clause, and
that a having clause with not group statements causes and error.