Adds a representation of a collection of columns
The column collection class can simplify selecting a common
subset of columns in a table multiple times, or a collection of
columns in two or more tables.

The column collection has also been used for columns that are not
strictly independent tables as far as a query is concerned, but
may need to be selected together, such as all columns in one table
of a joined table.