Bumps version to 0.1.1 and includes minor change
Column collections are the major change for this version increase.
The column collection itself helps to unify the columns of a table,
join, or query are managed internally in the package. It should also
be able to be used for storing combinations of columns for a number
of tables in a reusable way, though this has not been explored much.

Column collections present similarly to a mapping type, and are used
in a similar manner throughout the code that makes use of them.