Bumping version up to 0.2.1
This was going to wait until unit tests were written for the DML
functionality, but I am having difficulty figuring out how to mock
`sqlite3.Connection.commit` properly. The new DML functionality has
been tested, but this is an obstacle to writing good unit tests.

Adding a connection parameter to the `execute` method signature
for queries and DML was being considered as a modification, but will
probably be implemented sooner as a means to avoid this issue.