Adds an optional `conn` parameter to statement execute method
The argument should be a database connection, and allows
sharing of a connection among multiple SQL operations.
This can be convenient if several distinct statements need
to be treated as fully atomic.

This has also been useful in testing the DML classes. It has
been used to be able to query inserted, updated, or deleted
test data without actually commiting those operations.

When not passed in, this parameter defaults to None and the
statement will manage a connection itself, including committing
if relevant.