Defers the determination of DBAPI paramstyle
Using the new `breezeblocks.sql.Value` class, the way a
value renders into a query is not fixed at Value instantiation.
Instead it is determined when the statement is requested
from the statement builder.

This is a step towards a few potential enhancements, but is being
done in order to allow users to not need to explicitly construct
the classes from `breezeblocks.sql.values` or the `Value` class to
represent literals in the query.