Allows users to pass Python values to operators
When something that is not recognized by BreezeBlocks
as a SQL value expression is passed as an argument to an
operator constructor or statement builder, BreezeBlocks will
now make a value expression out of that. This is a usability
change to reduce the burden of using the `Value` class explicitly
because more often than not the intent is intuitive without it.