Adds first version of a parameter store
The parameter store represents the collection of statement parameters
passed to the DBAPI module. The impetus for adding this was to allow
the swapping of parameters for a pre-built statement so a common query
could be used multiple times. A simple use case is retrieving the same
fields of multiple records by changing the value of the id query
parameter.

A generic exception for any statement-building process is added here also.