Moves test logic into dedicated classes
The logic for test classes does not need to be on a unittest.TestCase
subclass. Inherited test methods will be run when discovered by the
unittest module. This allows us to write all test methods in once class
and extend another from it that does the correct setup for a DBAPI
and overrides methods for the tests that behave differently for that
SQL implementation.