Adds a getter and setter for value on bound parameters
I think it's a reasonable use case to want to get or set
the value of one of these that you have created and added to a
query, so this is being added as an accepted way of doing that.
These new methods are being called from the param store to access
or set the value also.