Fixes issue with INSERT INTO ... VALUES (...)
Insert into values was passing the data parameter,
which was expected to come in as a list of sequences,
to the dbapi module. Now that is being properly passed
through a param store first so that the dbapi gets the
parameters in the form it expects. This was affecting
mapped value dbapis, and probably numeric ones too.