Adds test modules for Postgres with Chinook database
Extending the base unit tests for DML and querying in a
class that sets up a Postgres database. As of now this
assumes that your server is running on localhost and takes
some connection parameters from the environment. The dbapi
module used is psycopg2.

Some tests do not pass yet for Postgres. Some that should work
have not been written and throw a NotImplementedError. Another two,
which are for ORDER BY functionality, seem to fail because the ordering
is case insensitive.