Adds a cursor proxy to support native cursors
In order to support DBAPI 2.0 modules that you cannot hold
a weakref to the cursor for, a cursor proxy is returned from
pooled connections which forwards mostly everything to the
underlying DBAPI cursor.
Context management capabilities were also added to the cursor
proxy that will close the cursor on exit of the with block.