Updates unit tests and improves debugging
Skips more tests for MariaDB which will not work specifically
for MariaDB. These are mostly due to limitation of ORDER BY in
MariaDB, and because it has no FULL JOIN.

The subquery in IN clause test has been improved such that it is
properly exercising that subquery and assertion messages have been
written.

`breezeblocks.sql.Query#show` prints a useful representation of
the statement parameters again.