Add an on_connect parameter to DB constructor
When provided, this should contain some SQL to execute every time
the connection is taken out of the pool. This is intended for
a statement like SQLite's `PRAGMA foreign_keys = ON`.

Additionally, this commit re-orders the DB constructor parameters
and makes some of them keyword-only. Most of the time probably
only dbapi_module and dsn should be necessary, so this is separates
what should be considered the "advanced" arguments.