Adds a `get_name` method to table expressions
This change seems pretty minor, but it's planned to be
useful in expanding the usability of joins by allowing named
access to the joined tables. Some types of table expression,
such as queries, there is not real obvious name, but this
provides access to what is considered the name and will be used
by the join also.