Adds a method for cloning a QueryBuilder
This is good for if you want to query the same fields
under different scenarios but change up the where clause.
No plans thus far to extend this to DML builders. I don't
think there is as much of a use case for those. It might make
sense for update, but I'm not sure yet.